HV-Pdf_Generator/requirements.txt
streamlit
python-docx
comtypes; platform_system == "Windows"
pythoncom; platform_system == "Windows"
Pillow
PyPDF2
pymupdf
num2words
streamlit
google-cloud-firestore
firebase-admin
google-cloud-storage
